[#8] refactor: DetailView를 DetailViewController와 분리
